fix: sum points in long before clamping
A forced award saturating at Integer.MAX_VALUE wrapped points negative,
so Math.max(0, ..) zeroed a player's bar while reporting the bar was
already full. The row then saved with points below claimed-points, and
the loader's clamp pulled claimed-points down to match, handing back
every milestone that player had already collected.

Also moves the points figure to its own message key, since a live
messages.yml keeps the old text for a key it already has.

+++ src/main/java/tfmc/justin/activity/models/PlayerData.java
@@ -201,8 +201,12 @@ public class PlayerData {
             awarded < earned ? Recorded.WEEKLY_CLAMPED : Recorded.RECORDED);
     }
 
+    // Widened to long before the sum: a forced add hands over an award that
+    // saturates at Integer.MAX_VALUE, and 'points + p' as an int would wrap
+    // negative and Math.max(0, ..) would then wipe the bar instead of filling
+    // it. Every value that did not overflow before is unaffected.
     public void addPoints(int p, int max) {
-        points = Math.max(0, Math.min(max, points + p));
+        points = (int) Math.max(0, Math.min(max, (long) points + p));
     }
 
     // Stored points can exceed the bar after bar.max is lowered or the file
